config: Rename oauth_config -> config.
